refactor: Moved some code to collector.go
* Moved some generic code to collector.go
* The "run collector" just prints help
